testCheckpointAtEndNotBuffered
